+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the "amiga-debug" extension will be documented in this fi
   - Syntax highlighting, symbol search and error squiggles support.
 - NEW: `DMA Control` shows `BltPri`
 - NEW: assembly: hover over custom register name to get documentation
+- FIX: screen: fix sprites over HAM
 
 ## 1.5.4
 - NEW: profiler: improve blitter tooltip for line mode, show simplified formulas of minterms